refactor: config, move methods
