+++ packages/config/package.json
@@ -25,8 +25,13 @@
   "dependencies": {
     "@verdaccio/dev-commons": "workspace:5.0.0-alpha.0",
     "@verdaccio/utils": "workspace:5.0.0-alpha.0",
-    "mkdirp": "0.5.5",
     "debug": "^4.2.0",
-    "lodash": "^4.17.20"
+    "js-yaml": "3.14.0",
+    "lodash": "^4.17.20",
+    "mkdirp": "0.5.5",
+    "yup": "^0.29.3"
+  },
+  "devDependencies": {
+    "@types/yup": "^0.29.9"
   }
 }
