+++ packages/config/src/index.ts
@@ -1,3 +1,5 @@
 export * from './config';
 export * from './config-path';
 export * from './token';
+export * from './config-utils';
+export * from './parse';
