+++ packages/config/test/config.spec.ts
@@ -3,8 +3,7 @@ import _ from 'lodash';
 
 import { DEFAULT_REGISTRY, DEFAULT_UPLINK, ROLES, WEB_TITLE } from '@verdaccio/dev-commons';
 
-import { parseConfigFile } from '@verdaccio/utils';
-import { Config } from '../src';
+import { Config, parseConfigFile } from '../src';
 
 const resolveConf = (conf) => {
   const { name, ext } = path.parse(conf);
