+++ packages/config/test/partials/config/js/default.js
