+++ packages/config/test/partials/config/js/invalid.js
