+++ packages/config/test/partials/config/json/README.md
