+++ packages/config/test/partials/config/json/default.json
