+++ packages/config/test/partials/config/json/invalid.json
