+++ packages/config/test/partials/config/yaml/deprecated-pkgs-basic.yaml
