+++ packages/config/test/partials/config/yaml/pkgs-basic-no-access.yaml
