+++ packages/config/test/partials/config/yaml/pkgs-basic-no-proxy.yaml
