+++ packages/config/test/partials/config/yaml/pkgs-basic-no-publish.yaml
