+++ packages/config/test/partials/config/yaml/pkgs-basic.yaml
