+++ packages/config/test/partials/config/yaml/pkgs-custom.yaml
