+++ packages/config/test/partials/config/yaml/pkgs-empty.yaml
