+++ packages/config/test/partials/config/yaml/pkgs-multi-group.yaml
