+++ packages/config/test/partials/config/yaml/pkgs-nosuper-wildcard-custom.yaml
