+++ packages/config/test/partials/config/yaml/uplink-basic.yaml
