+++ packages/config/test/partials/config/yaml/uplink-wrong.yaml
