+++ packages/mock/package.json
@@ -25,6 +25,7 @@
   "dependencies": {
     "@verdaccio/dev-commons": "workspace:5.0.0-alpha.0",
     "@verdaccio/utils": "workspace:5.0.0-alpha.0",
+    "@verdaccio/config": "workspace:5.0.0-alpha.0",
     "fs-extra": "^8.1.0",
     "lodash": "^4.17.20",
     "request": "2.87.0",
