+++ packages/utils/src/index.ts
@@ -2,5 +2,4 @@ export * from './auth-utils';
 export * from './string';
 export * from './utils';
 export * from './crypto-utils';
-export * from './config-utils';
 export * from './replace-lodash';
